requirements.txt
autopep8==1.4.4
dataclasses==0.6
dj-static==0.0.6
Django==2.2.13
djangorestframework==3.10.3
djongo==1.2.36
dnspython==1.16.0
Faker==2.0.2
pycodestyle==2.5.0
pymongo==3.9.0
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.2.4
static3==0.7.0
text-unidecode==1.3
